Adding more CSS
Likely won't finish CSS, would be able to with more time.
