+++ src/components/Body/Body.js
@@ -1,7 +1,7 @@
 import React from "react";
 import GetStarted from "../GetStarted/GetStarted";
-import InfoGrid from "../InfoGrid";
-import Testimonials from "../Testimonials";
+import InfoGrid from "../InfoGrid/InfoGrid";
+import Testimonials from "../Testimonials/Testimonials";
 import StayProd from "../StayProd/StayProd";
 import SignUp from "../SignUp";
 import CompInfo from "../CompInfo";
