+++ src/components/Header/Header.css
@@ -1,8 +1,9 @@
+* {
+    font-family: Open Sans;
+    background-color: hsl(217, 28%, 15%);
+}
+
 @media only screen and (max-width: 375px) {
-   
-    * {
-        font-family: Open Sans;
-    }
 
     .logo__left {
         margin: 3vw;
