+++ src/components/InfoGrid/InfoGrid.css
@@ -1,0 +1,7 @@
+.infoContainer {
+    margin-top: 70px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    text-align: center;
+}
