+++ src/components/Testimonials/Testimonials.css
@@ -1,0 +1,3 @@
+.testBox {
+    background-color: hsl(219, 30%, 18%);
+}
