+++ src/index.css
@@ -7,6 +7,7 @@
    
     * {
         background-image: url("./images/bg-curvy-mobile.svg");
+        background-repeat: no-repeat;
     }
 
 }
@@ -15,6 +16,7 @@
 
     * {
         background-image: url("./images/bg-curvy-desktop.svg");
+        background-repeat: no-repeat;
     }
 
 }
